pyproject.toml
[project]
name = "mas-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "keyboard>=0.13.5",
    "numpy>=2.4.1",
    "openai>=2.14.0",
    "pydantic[email]>=2.12.5",
    "pynput>=1.8.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sounddevice>=0.5.3",
    "soundfile>=0.13.1",
]
